backend/requirements.txt
# FastAPI stack
fastapi[standard]==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# DB
sqlalchemy[asyncio]>=2.0.40
asyncpg==0.30.0
alembic==1.14.0
geoalchemy2==0.16.0

# Auth & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# passlib 1.7.4 expects bcrypt<4.1 (no __about__ in bcrypt 4.1+)
bcrypt>=4.0.1,<4.1
setuptools>=65,<78
slowapi==0.1.9
authlib==1.4.0

# Redis & Celery
redis==5.2.1
celery[redis]==5.4.0

# HTTP / OAuth / Payments
httpx==0.28.1
stripe==11.4.1
razorpay==1.4.2

# Config
pydantic-settings==2.6.1
email-validator==2.2.0

# Geocoding (optional signup)
# Nominatim via httpx — no extra package
